ai/requirements.txt
transformers==4.35.0
torch==2.0.0
flask==2.3.0
numpy==1.23.5
scikit-learn==1.3.0
nltk==3.8.0
flask-cors==4.0.1
python-dotenv==1.0.0
requests==2.31.0
playwright==1.47.0
Pillow==10.4.0